8286394: Address possibly lossy conversions in jdk.naming.dns

+++ src/jdk.naming.dns/share/classes/com/sun/jndi/dns/DnsClient.java
@@ -730,7 +730,7 @@ class Tcp {
             return reader.read();
         }
         finally {
-            timeoutLeft -= System.currentTimeMillis() - start;
+            timeoutLeft -= (int) (System.currentTimeMillis() - start);
         }
     }
